siaknng/auth/requirements.txt
django==2.2.7
djangorestframework==3.10.3
django-rest-auth==0.9.5
dj-database-url==0.5.0
gunicorn==20.0.4
